postprocess: Symlink /etc/resolv.conf to /run/NetworkManager if NM
We want `/etc/resolv.conf` to be a symlink to `/run` so that one
can reasonably bind mount read-only `/etc`.  This sadly isn't the
default...epic series of bugs here; see e.g.
https://bugzilla.redhat.com/show_bug.cgi?id=1197204

Particularly for Fedora CoreOS with Ignition in play where the mental
model is "configure once", supporting read-only `/etc` is nice.

This is the first content-dependent change we're making; it's better
to have it here than in everyone's postprocess-script.