compose: adds integration parts to compose
The logic for writing sysusers is integrated into both compose logic
and rpmostree_check_passwd_groups.

The plan is that whenever "sysuser" is specified in check-passwd section,
The compose will use the sysuser logic. The passwd entries instead will
be converted to sysuser entries. And eventually at the end of compose
a file will be written to '/usr/lib/sysusers.d'.