status: Remove rpmostree.rpmdb.pkglist from JSON
Sometimes I eyeball the json to get some data, and yeah, probably
should be using a toolbox container and jq, but still it's
convenient.

The addition of the pkglist to the JSON (since the JSON includes
all commit metadata) made that totally illegible;
this also came up with atomic-host-tests which was using it.

I don't forsee anyone adding anything even remotely as
large as the pkglist to commit metadata, so let's just strategically
nuke it.