compose: Add a helper to convert treefile → treespec
At some point perhaps we'll define a sane format that unifies
treefile/treespec.  This is not that day, but let's extract
the code that converts the two into a clean helper.  Then the
compose context holds onto the treespec, and e.g. things like the
ref just point to its data.

Prep for sharing this code with rojig.