app: Wrap `Version` in `--version` in quotes
Wrap the `Version` key in the YAML-compatible output of
`rpm-ostree --version` with quotes so that it's parsed as a string. The
issues with the previous approach in a nutshell:

```
In [5]: yaml.load("asdf: 2018.10")
Out[5]: {'asdf': 2018.1}
```

It's treating the version number as a floating-point. Now, this is
technically a backwards incompatible change, but given that the previous
approach is inherently broken for our needs, I don't see a way around
breaking it now.

See also: https://github.com/ostreedev/ostree/commit/759b099861a3065f01384d4d1121a7726771e5fe

Closes: #1654
Approved by: cgwalters