app/compose: Add --legacy to start deprecation process
First, start printing the current mode we're operating in. This will
make it clear from the output if the default changes. Second, add a new
`--legacy` switch that users can use today to explicit opt into the
current default. And third, start emitting a warning if neither
`--legacy` nor `--unified-core` is given.

The way I'm thinking we can do this is (timespans flexible of course):
- next 6 months: print warning unless mode is explicitly chosen
- following 6 months: change default, print deprecation warning if
  `--legacy` is given
- afterwards: remove `--legacy`, gradually clean up code, and profit